feat(x/gov): constitution amendments can update the constitution (#25)
The `ProposeConstitutionAmendment` msg that is used for constitution amendment proposals now contains an `amendment` field, that needs to be in *unified diff* format (the output of `diff -u`) and will trigger an update
of the constitution.

The choice of using a unified diff format instead of requiring to overwrite the constitution with an entirely new string is for a couple reasons:

- human readability
- compactness

The choice of the unified diff format was deemed the best tradeoff to achieve both results.

A cli command in the txs called `generate-constitution-amendment` is also provided that fetches the current constitution and generates the amendment message that needs to go in the proposal, but the patching mechanism accepts standard unified diffs generated from any source. Putting it into the txs seemed the best place.

The code that checks and applies the diffs does not rely on any external library, conversely the CLI utility to generate the diff does make use of `github.com/hexops/gotextdiff`.