docs(adr): remove wrong statement
The current TxFeeChecker does not check against *all* the denom listed
in the validator minimum-gas-prices, but against *at least* one of the
denom listed. So actually by listing both uatone and uphoton, we can use
the same check in the photon TxFeeChecker, it will work whether photon
is required or not.

See `feeCoins.IsAnyGTE()` function.