feat(x/photon): validate fee ante decorator (#54)
Relates to #44 

This PR adds a `sdk.AnteDecorator` implementation in `x/photon/ante` and
adds it to the list of the app ante decorators. This new ante decorator
ensures that photon is the fee token or returns an error. Exceptions are
possible like if the transaction's messages type URL are present in a
new parameter of the photon module `txFeeExceptions`.

Some significant changes have been made to the e2e tests to make them
compatible with this conditions.

`uatone` has also been put in a constant to avoid repetition (in
`app/params` package)