chore: bump go to 1.22
Go is being  carefully upgraded in SDK chains because of issues with
state sync between nodes (although there is no evidence of this as far
as I have investigated).

This change upgrades Go from 1.21.13 to 1.22.10 (the latest 1.22
release). Go 1.22 is already being used in the Gaia mainnet and no
problems of state sync have been reported.

Once merged, this will allow the upgrade of cometbft which will fix 2
vulnerabilities raised by govuln.