chore: remove `x/crisis` module (#93)
This PR removes `x/crisis` because it has been identified as a security
issue.

Originally, the module could be used by anyone to halt the chain in the
event of a violated invariant. To prevent spamming the requested would
be charged a `ConstantFee` .

However, the module has been shown to **NOT** [halt the
chain](https://github.com/cosmos/cosmos-sdk/security/advisories/GHSA-qfc5-6r3j-jj22)
and to [not
charge](https://github.com/cosmos/cosmos-sdk/security/advisories/GHSA-w5w5-2882-47pc)
the `ConstantFee` to the requester.

This implies that an attacker could have used `x/crisis` as a vector for
a DOS attack.

---------

Co-authored-by: Giuseppe Natale <12249307+giunatale@users.noreply.github.com>