fix(x/gov): proposal v1 to v1beta1 converter (#102)
Converter was returning an error when the v1 proposal has no messages,
which can be considered as a bug since the v1 API allows to submit
proposals without messages.

The fix handles such proposal and turn them into v1beta1 text proposal.

Additionnaly to unit tests, a test was made with a local node, ensuring
it is now possible to request a v1 proposal that has no messages using a
v1beta1 request.

---------

Co-authored-by: Giuseppe Natale <12249307+giunatale@users.noreply.github.com>