test: fix tx priority computation
tx priority was using the denom resolver to get the fee value in
uphoton, when the fee sent in the tx use a different denom. This was not
working because the denom resolver can only convert uphoton to uatone,
and not the other way around.

Fixed by not calling the denom resolver but rather computing the
conversion factor that was used for the initial conversion, and using
that factor to convert back the amount of fee to uphoton.

Also reverted the removal of expected keepers in post and ante package.