refactor(x/feemarket): various improvements  (#113)
- register new module migration
- enable AIMD mode in default genesis
- rename baseFee to baseGasPrice (in code and docs)
- register new error for max block size exceeded
- improve some tests
- adjust AIMD default gas price

This last change breaks the test case
`TestState_UpdateBaseGasPrice/half_target_block_size_with_aimd_eip1559_with_a_delta`
and I found difficult to fix it. The problem is related to the usage of
the `delta` parameter, and I think we should remove it for several
reasons:

- `delta` is not used (=0) and we have no plan to use it in the future
- `delta` constant is absent of the original paper
https://arxiv.org/pdf/2110.04753
- `delta` is a multiplier of the net gas difference between every block
in the window and the target block size, so the result is expressed in
gas, and then added to the gas price. Gas and Gas price have 2 very
different order of magnitude, the first has a maximum of `30,000,000`
while the other is usually a small decimal number like `0.01`. Unless I
missed something that makes no sense to add those 2 numbers. It is
revealed now because this PR updates the base gas price of AIMD params
from `1,000,000,000` to `0.01`.

Please share me your thoughts guys.

EDIT:

It has been decided to remove the `delta` parameter. The motivation for
this is that Delta, when enabled, does not provide a behaviour that we
would like to have, and if necessary we think it is possible to
reproduce its behaviour with the remaining parameters.

Removing it also simplify the code and remove extra computations.