docs: Adding ADR-004, Nakamoto Bonus (#127)
## Description

Nakamoto bonus ADR (004)
This ADR proposes a mechanism to increase the Nakamoto coefficient of a
chain, by incentivizing delegators to delegate on validators having less
stake. The ADR proposes to divide the staking reward in two components:
a part proportional to the overall amount of stake of a validator - this
coincides with the current staking reward model, and a part that is
distributed uniformly to validators - the `Nakamoto bonus`.
The overall amount of the reward remains unchanged, however the criteria
for assigning the reward changes.

[Rendered
version](https://github.com/atomone-hub/atomone/blob/giuliostramondo/docs/adr-014-nakamoto-bonus/docs/architecture/adr-004-nakamoto-bonus.md).

---------

Co-authored-by: Thomas Bruyelle <thomas.bruyelle@tendermint.com>