feat(x/gov): dynamic deposit increases only with proposal activation, and decreases only with time (#105)
This PR is based on #104 

Based on discussions, this is a revised model for the dynamic deposit
that:
- only performs deposit increases upon proposal activation (proposal
deactivation does not trigger an update of the deposit), and
- only perform time-dependent updates to decrease the deposit, and a
decrease happens only if active proposals is less than the target,
otherwise the time-based update has no effect.

Plus, remove the sensitivity to the distance for increases.

This PR should be merged only after #104 is (or in its stead). Moreover,
if this PR is accepted, the partially overlapping PR #103 can be closed.

---

Importantly, this PR does not yet update neither the `x/gov` README.md
nor the Dynamic Deposit ADR to reflect these changes, so it needs to be
done.

---------

Co-authored-by: Thomas Bruyelle <thomas.bruyelle@tendermint.com>