test(e2e): fix feemarket price change
An error in the genesis file was causing the feemarket price change test to
fail, because the set of the max block gas was ignored. The default
value of -1 was causing a max block gas being equal to maxUint64 in the
feemarket module, making the target block gas also very high and so the
required amount of gas to see an increase of the gas price was not
possible.

The error was caused by the usage of cometbft/libs/json to marshal the
genesis in the e2e tests, bc since 0.50, the genesis file is read using
the legacy encoding/json, and those libs does not treat integer the same
way (cometbft/libs/json marshals integers into string).
So the fix is to replace cometbft/libs/json  with the legacy encoding/json
in the e2e setup.

Also had to increase the max block gas in the e2e tests from 50,000,000 to
60,000,000 because the previous max was reached in the
`testFeemarketGasPriceChange` test.