fix(gov/simulation): return zero if max-min <= 0 for certain generated params (#168)
The failed run for PR #167 for the nondeterminism test (see
https://github.com/atomone-hub/atomone/actions/runs/16119695430/job/45482097081?pr=167)

made me realize we need to avoid to have `n <= 0` for `Intn()` (called
by `RandIntBetween`) since this triggers a panic.
This can happen if max ends up being == min (== 0) which may very well
happen for certain generated genesis parameters in simulation that have
parametric maximum.

Therefore, I added an if statement that returns a `sdk.ZeroDec` value in
case this happen, avoiding failure.