fix(e2e): E2E tests do not fetch dynamic deposit values (#196)
The dynamic deposit feature makes the required minimum initial deposit
and overall deposit dynamic. E2E tests use for both a fixed amount, this
can cause the tests to fail if the fixed amounts used are less than the
dynamic amounts required.
 
This PR adds queries for the dynamic deposit values and uses the updated
values for the e2e tests.