docs: generate key from custom entropy
Similarly to `gnokey add mykey --entropy` [0], this section describes
how to generate a key from a custom user entropy.

Since there's already a command to generate a mnemonic from a custom
entropy (`keys mnemonic --unsafe-entropy`), I deemed unecessary to
add a `--entropy` flag to the `keys add` command like in `gnokey`. As a
result, generating a key with a custom entropy is a 2-step process, but
that's probably good enough.

[0]: https://github.com/gnolang/gno/pull/4586